Overhaul USDC writer, fix roundtrip gaps, add comprehensive tests
USDC writer path encoding:
- Replace EncodePaths-based path tree with direct pxrUSD-style
  BuildPathIndicesFromSortedPaths using Path::operator< sorting and
  pre-assigned PathIndex values (no intermediate UINT64_MAX sentinels)
- Add ";-)" sentinel at token index 0 matching OpenUSD workaround for
  github issue #811 (-0 == 0 property path ambiguity)
- Fix NurbsCurves roundtrip: add GeomNurbsCurves to GetPrimElementName
  and SetPrimElementName dispatch tables in prim-types.cc
- Fix Path::_update for relative paths: set _element for ndots==0 case
- Add "properties" field to prim specs for Pixar reader compatibility
- Add "primChildren" to pseudoroot spec

USDC writer property gaps fixed:
- Camera: guard all properties with authored() to prevent default injection
- Mesh: fix subdivisionScheme/interpolateBoundary/faceVaryingLinearInterpolation
  default injection; add velocities extraction; add normals interpolation
  metadata; route doubleSided through ConvertAttributeToFields
- Points: add ids (int64[]) extraction
- Skeleton: add bindTransforms/restTransforms (matrix4d[]) extraction
- SkelAnimation: add scales (half3[]) extraction
- All lights: add shadow API, shaping API, common light properties;
  fix inputs: prefix; fix has_value()->authored() default injection;
  fix DomeLight file asset type
- Shader: add terminal output handling for UsdUVTexture, UsdTransform2d,
  UsdPrimvarReader_* with correct type names
- Material: add MaterialXConfigAPI property extraction
- Fix ConvertPropertyToFields EmptyAttrib handling for terminal attrs

Crate writer type support:
- Add WriteValueData for matrix2d/3d/4d[], half2/3/4[], int2-4[], uint2-4[]
- Add ConvertValue handlers for texCoord2f/2d/3f/3d[], float2[], half2-4[],
  int2-4[], bool[], matrix2d/3d[]
- Remove matrix[]/half[] upstream filters now that types are supported
- Add GetPrimProps entries for NurbsCurves, PointInstancer, NodeGraph,
  GeometryLight, PortalLight

Parser and pprint fixes:
- USDA parser: reject multiple property statements on single line
- Fix wrapS/wrapT name swap in pprint-shader.cc
- Fix default value injection for connected token attrs in pprint-detail.hh

Tools and testing:
- Rewrite tusddiff to use LoadLayerFromFile with proper exit codes
- Add usdc-writer-runner.py (tusdcat -> USDC -> tusddiff pipeline)
- Register usdc-writer-diff-test in CMake
- Add 28 comprehensive USDC writer unit tests covering usdGeom, usdSkel,
  usdShade, usdLux, usdMtlx, instancing, and apiSchemas
- Update doc/testing-cpp.md with new test coverage

+++ src/value-types.hh
@@ -2316,7 +2316,10 @@ class ValueView {
   };
 
   // Default constructor - creates an invalid view
-  ValueView() noexcept : ptr_(nullptr), type_id_(TYPE_ID_INVALID), flags_(FLAG_NONE) {
+  ValueView() noexcept
+    : ptr_(nullptr),
+      type_id_(static_cast<uint32_t>(TYPE_ID_INVALID)),
+      flags_(FLAG_NONE) {
     std::memset(padding_, 0, sizeof(padding_));
   }
 
@@ -2335,7 +2338,8 @@ class ValueView {
   // Constructor from const Value pointer
   explicit ValueView(const Value* value) noexcept
     : ptr_(value),
-      type_id_(value ? value->type_id() : TYPE_ID_INVALID),
+      type_id_(value ? value->type_id()
+                     : static_cast<uint32_t>(TYPE_ID_INVALID)),
       flags_(FLAG_NONE) {
     std::memset(padding_, 0, sizeof(padding_));
     if (value && (type_id_ & TYPE_ID_1D_ARRAY_BIT)) {
@@ -2543,7 +2547,7 @@ class ValueView {
   // Reset the view
   void reset() noexcept {
     ptr_ = nullptr;
-    type_id_ = TYPE_ID_INVALID;
+    type_id_ = static_cast<uint32_t>(TYPE_ID_INVALID);
     flags_ = FLAG_NONE;
   }
